Authentication implementation
SEE /backend/README.md API/Authentication for details or /backend-testing/index.js for examples. Usernames and passwords should be supplied on the user post route but are hidden on the get routes. Passwords in the database are hashed, but passwords sent to the system are in plaintext, idk how to fix that or if that's even something we should be worried about.